requirements.txt
Django==2.1.5
djongo==1.3.1
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
constants==0.6.0
